changed divide_factor to float. Divided the out_dim of pe_encoders.

+++ goli/ipu/ipu_simple_lightning.py
@@ -45,7 +45,7 @@ class SimpleTorchModel(torch.nn.Module):
             nn.LogSoftmax(1),
         )
 
-    def make_mup_base_kwargs(self, divide_factor: int = 2):
+    def make_mup_base_kwargs(self, divide_factor: float = 2.):
 	    return dict(
             in_dim=self.in_dim,
 	        hidden_dim=round(self.hidden_dim / divide_factor),
